refactor(userbodymetrics): reorganize packages with port package
