+++ src/main/java/com/ironcore/application/userbodymetrics/list/ListUserBodyMetricsUseCase.java
@@ -4,6 +4,7 @@ import com.ironcore.application.exception.ResourceNotFoundException;
 import com.ironcore.application.exception.UserInactiveException;
 import com.ironcore.application.shared.pagination.PageQuery;
 import com.ironcore.application.shared.pagination.PageResult;
+import com.ironcore.application.userbodymetrics.port.ListUserBodyMetricsQueryPort;
 import com.ironcore.domain.user.model.User;
 import com.ironcore.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
