+++ src/main/java/com/ironcore/infrastructure/persistence/userbodymetrics/repository/ListUserBodyMetricsQueryAdapter.java
@@ -3,7 +3,7 @@ package com.ironcore.infrastructure.persistence.userbodymetrics.repository;
 import com.ironcore.application.shared.pagination.PageQuery;
 import com.ironcore.application.shared.pagination.PageResult;
 import com.ironcore.application.userbodymetrics.list.ListUserBodyMetricsItemResult;
-import com.ironcore.application.userbodymetrics.list.ListUserBodyMetricsQueryPort;
+import com.ironcore.application.userbodymetrics.port.ListUserBodyMetricsQueryPort;
 import com.ironcore.domain.user.valueobject.UserId;
 import com.ironcore.domain.userbodymetrics.valueobject.BodyHeightCm;
 import com.ironcore.domain.userbodymetrics.valueobject.BodyWeightKg;
